+++ src/abstractclipitem.h
@@ -68,7 +68,7 @@ protected:
     GenTime m_cropDuration;
     GenTime m_startPos;
     GenTime m_maxDuration;
-    QMap <int, double> m_keyframes;
+    QMap <int, int> m_keyframes;
     double m_keyframeFactor;
     double m_keyframeDefault;
     double m_fps;
